Soft wrap at 72 chars in mutt
This works by limiting the total number of columns when composing an
email with mutt. Instead of using 'textwidth', I use 'columns' to limit
the number of columns on a screen. This gives me a shorter line length
without the line breaks that 'textwidth' would introduce.

The reason I use this approach is because GMail does not play nice with
format=flowed, so it's easier to just have long lines in plaintext email
and have the recipient's email client wrap them automatically.