Use termite as main tty
Alacritty has been a little buggy for me lately (not displaying Chinese
characters properly, not renewing when an animated CLI/TUI process is
running). I tried kitty first, it's pretty nice but not compatible with
fcitx, and the maintainer stated he had no plans to write support for it
himself. Termite is nice, easily configurable, and works for all my needs.