Change org external app associations
I used to do this through a custom function, but then I found that it's
just as easily done by directly modifying the `org-file-apps` variable.
I have to wrap it in an `after!` macro though, probably due to lazy
loading in Doom.

Also cleaned up unnecessary cruft and changed the relevant keybindings.

Still can't get org to use qutebrowser instead of chromium tho, will
have to deal with that later.