Move plugin block near top of .vimrc
This change was mainly due to a colour scheme ('darcula') being a
plugin, but nvim was complaining when the plugin was further in the file
than the line `colorscheme darcula`. I guess ordering is
important in VimScript. Best move all plugins up in case this issue
comes up again in the future.

Also added some extra plugins.