doom: add lisp keybinds
Added new keybindings for slurping and barfing in either direction. The
bindings work in both Elisp mode and in Racket mode, and it will be
trivial to add more Lisp modes to the list if need be.

Slurp and barf come from smartparens-mode, and are great for editing
Lisp code.