emacs: renew init.el
This file does not get updated with Doom, but it does get changed as new
packages come in and old ones get dropped. I thought now would be a good
time to pull in a fresh example init.el and add the packages I'm using.

This let me see what other new stuff hlissner has added, as well as try
ivy over helm since it's apparently the default now.

The layout of the file has changed somewhat, and the comment block in
the beginning is new, so that makes this diff especially big.