emacs: remove cnfonts package from config
Finally found the culprit behind my font shenanigans. It was the hook in
the cnfonts package messing with Doom's fonts. I tried disabling the
package and the problem went away.

The only tradeoff seems to be that I won't have a tool to adjust font
sizes interactively, but I can live with that. I don't think I need
perfect alignment in my Chinese fonts, but even if I do one day, I can
do it manually.