emacs: insert pandoc-style cites in org-mode
When exporting with `ox-pandoc`, org-ref citations don't work for
multiple cites in the same parentheses, at least for me. Pandoc-style
citations do work as usual, so I will stick to these for now.

However, the `bibtex-completion-format-citation-pandoc-citeproc`
function inserts square brackets around every new cite that I call,
which is annoying for two reasons:

1. I don't always want parenthetical citations,
2. I don't need parentheses if I already have a pair and I'm inserting
additional cites inside them.

This function, and all functions in
`bibtex-completion-format-citation-functions`, take a list of
strings (keys) as their only parameter. Problem is, I never pass a list
to these functions, but only one key at a time (I'm not sure what the
original intended idea was).

The `org-ref-format-citation` function checks if the point is on an
org-ref citation, and if it is, appends to it instead. This seems like a
workable solution, and I could try and implement something like it.