emacs: fix source blocks in ox-pandoc
Source blocks do not work out of the box in ox-pandoc because it strips
the source tags away. In order for them to work, they need to be wrapped
in a second source block, however this is clunky and affects syntax
highlighting. Instead I use a hook that does this automatically during
export.

The information, function, and hook are all taken from here:
https://github.com/kawabata/ox-pandoc/issues/51