rofi: change prompt
Removed the unicode symbol that I don't even remember what it was, but
it's not in the font I'm currently using with rofi.
Replacing it with a colon works just as well, since I don't even notice
the prompt most of the time (I do, however, notice when there's a
missing symbol).

Also increased the font size by the by.