mu4e: check for recipient field before sending
So I have a habit of not filling out the recipient field until the
message is written, from a (probably unfounded) fear that I will
accidentally send an unfinished message.

Sometimes what I end up doing instead is trying to send the finished
message to nothing, because I forgot to add the recipient's address.
This is a minor problem, exacerbated by org-msg, which transforms the
text of the message into html *before* trying to send it. If it fails,
then I get a bunch of html in the email which is visible to the
recipient.

This fixes the problem by refusing to send if there is no one to send
to. Perhaps this should be fixed upstream. Note to self: file a bug
report later.