Set DPI scaling to 125%
I tried 150% for a while, and while the text was nice and large on my
laptop screen, it also enlarged the icons in the tools tray of GUI
applications too much, making them ugly and taking up too much space.

125% is the sweet spot, where things that need to be scaled are just big
enough that I don't have to strain my eyes, but not too big to waste
screen real estate.

I had to adjust the text size in the terminal accordingly.