Disable pop-up windows by default
One thing that is annoying when using Emacs with a tiling window
manager, is trying to manage (Emacs) windows inside a single Emacs frame.
I'd rather not do it at all and rely on frames instead. However, some
commands open a pop-up window by default, and this is also part of
Doom's config. I have found a way to disable this by default, but it was
also trivial to implement a toggle keybinding for this functionality if
I feel I need it.

Note that this makes Emacs info pages and other things like eww buffers
open in a new buffer in the same window (instead of a new window in the
same frame)[1], but still opens help for functions/variables/modes in a
popup window. This behaviour is controlled by the `pop-up-windows`
variable, which is `t` by default. I prefer to have it on, as function
docs are just fine in a small popup window, but I want info mode in
fullscreen.

While I was adding the new keybinding, I also sharp-quoted the function
pointers for all of my global keybindings. See the following for why:
https://endlessparentheses.com/get-in-the-habit-of-using-sharp-quote.html

[1]: Note that `window` in Emacs means something different from how the
word is normally used in operating systems. If you are reading this and
your name is Matt, the tl;dr is: OS windows are Emacs frames, and Emacs
windows are 'splits'. An Emacs buffer is any open file or Emacs program.