Remove nyan mode
Nyan-cat mode ~~is for pussies~~ isn't helpful when I want to tell at a
glance where on a page I am, especially with longer pages. It was fun,
but it's time to end this.

Another reason is that it uses a hacky solution, and qutebrowser will
include proper support for plugins in the future, so it will become
obsolete soon anyway.