Center screen after searches with / and ?
I have screen centering after n and N, this makes for more consistent
behaviour. I was originally worried that this might make for an awkward
experience, but with incremental highlighting it's not jarring at all,
and the centering happens only when I press Enter.
I needn't have worried.

As an aside, this change required only adding the function names to my
center-screen-after macro invocation. So cool to have a nice macro
interface.