Ebib: fix year retrieval in gd/ebib-get-year
Before the fix, the NOERROR slot of the `ebib-get-field-value` function
was a string. This wasn't ideal, because if `ebib-get-field-value` can't
find a value, and NOERROR is a string, it returns that string instead.
This was especially problematic for entries with an empty `year` slot,
but that still had the year in `date`. Changing it to a symbol fixes the
issue.