mu4e: change mail retrieval command
offlineimap is written in Python 2, which has now officially passed its
end of life. It's being rewritten in Python 3, but that is a work in
progress. Similarly, getmail is also Python 2 (there is a Python 3 fork,
but unlike the offlineimap rewrite, this one is not official). I've
considered mbsync, but it only supports IMAP, and I need POP3 too. I
decided to go with fdm right now. It's not being actively developed
these days, but it's a simple program, so it won't stop working any time
soon.

Gotta be careful to tell it to keep the mail on the server though.
Little bastard nuked both my mailboxes.