Rehaul mail config
It's still a work in progress, but I did manage to get some variables
out into a context. The main problem is that I can't use the NCUE
address to send mail when I'm not physically there, but perhaps I'll add
more contexts in the future.

Also fixed the bookmarks to only show inbox messages.