mu4e: disable visual-line-mode in header view
This allows me to have one email per line. The email subject will often
go beyond the screen edge, but I can just open the email to read it. The
plus side is that threads are much more salient, and I can see more
emails on the screen.

I did a similar thing with ebib in the past.