org: make agenda always start on a Monday
This is a somewhat hacky solution, maybe the manual has a better way of
doing the same thing. Right now if I open the agenda on a Monday, it
drops me into the previous week. Although that may not be too much of a
hassle, since I can just press a single button to go forward. A lot less
annoying than not having the week start on Mondays each time.