org: always start the agenda on this week's Monday
This can't be done with org's in-built "-Mon", because on a Monday it
starts the agenda on the previous week instead. In order to always start
it on the current week, I used a hook to calculate the start day when
starting the agenda, based on what day of the week it is today.