Make evil operate on logical lines
With visual-line-mode, and evil-respect-visual-line-mode enabled, evil
will do linewise operations (such as D) on the visual line, and not the
logical lines of text in the file. I don't want that. Moreover, that's
not the behaviour I'm used to. I think it's changed before in Doom, but
I'm not sure. Anyway, I have now found the culprit and pinned it.