Add new function `advice-add!`
In Doom's spirit, this function wraps around `advice-add`, but allows
both symbols and functions to be lists. Lifted almost verbatim from
general-advice-add, but I did learn a thing or two (and rid myself of a
possible dependency issue down the line).

I also rewrote my initial call to a less general macro with this new
function.