racket-mode: modify indentation in `do` blocks
Racket already has a `do` function, but I mostly use the one from
data/monad (together with parser combinators), so I need a different
kind of indentation.

Modifying it turned out to be pretty simple:
https://www.racket-mode.com/#racket_002dindent_002dline
However I had to make it a hook instead of a line in my config.