Add function to scroll line to just below top
I don't like scrolling the current line to the very top, it's a bit
difficult to read from the very top of the screen. I leave two lines
above it, and that works great! Made it into a function and replaced
previous scroll to top advice with this new one.