ebib: allow adding files to papers from index view
Normally the import file function in the index mode will instead create
a new entry. What I wanted instead was have a function that would add an
attachment to the entry at point. I wrote one, and gave it the same
keybinding I use in the entry mode, so now I can add attachments without
worrying which mode I'm in.