org-moar: prefill note title on link insert
The link insert function asks for the text to insert together with the
link. In cases where I want to use the original note title or something
close to it, it would be a hassle to type out the whole note. For cases
where I want to insert something else entirely, I only need to press C-u
to delete the prefilled text.