Don't defer org-books keybind
Having an after block prevents the keybind from being assigned until
after org-books has loaded. However, this is the exact opposite of what
I want in a fresh Emacs session: I want to be able to open the books
file from anywhere, loading org-books in the process if necessary. Since
org-books-jump-to-reading is autoloaded, it will load the rest of the
mode when called.