XCompose: add key combos for lax vowels
This required some restructuring.

I had ʉ and ɨ on u+u and i+i respectively, but I wanted the lax vowels
on these, so I needed to get the barred ones out of the way. The combos
vowel+- or -+vowel wouldn't work, since those are taken by ū and ī, and
I don't want to change them because all other vowels are already set up
this way.

So I settled on <-> <-> <vowel> for now, but I might change it later.