org-mode: add phonemic transcription snippet
Simple slashes give italics in org-mode, and escaping the slashes
actually produces a \/ sequence when compiling a PDF with pandoc (not to
mention it's a pain in the arse to type).
This snippet is simpler and less of an eyesore, but it works only for
LaTeX output. Which is fine at the moment.