org-moar: use abbreviated path in links
My use case is that my username is different on different machines, but
the rest of the path relative to $HOME is the same. Using `~` for $HOME
works here, but if I use the long file path instead, then it inserts the
username on the current machine, making the link unusable on other
computers.