org: add new link-with-id fn
This function adds a link to a heading in the buffer, but creates a
unique ID and binds the link using that (if an ID already exists, the
function reuses it).

This is helpful in two cases:
1. If the heading ever changes, the links will still point to the
correct place.
2. If there is more than one heading with the same name, links based on
heading names will only ever find the first one. ID based links will
always point to the correct heading.

Also added a keybinding for the function.