Show properties when cycling in org-books
By default (Doom's, I think), cycling through heading visibility in
org-mode using TAB never shows the PROPERTIES drawer. It has to be
cycled on its own instead. This is fine for most org-mode files, but in
org-books-mode the PROPERTIES drawer contains information that I want
visible at all times.

I implemented this via removing the function from the org-cycle-hook
locally in org-books-mode. The problem was that a buffer-local
org-cycle-hook doesn't exist by default. So I create it manually and
modify it with a setup function, which I then call when the mode is
activated (via a mode hook).

This lets me keep cycling behaviour different between org-books-mode and
normal org-mode buffers.