mu4e: remember location of last local attachment
Bind ", a" to a new function that sets the local value of
`default-directory' to the location of the last attachment. This value
is used by `read-file-name' as the starting location of the search.
A little quality of life improvement that remembers where I left off
with the previous attachment, so I can more easily attach several files
from a single directory at once. The value of `default-directory' is
buffer-local, so this is reset with every email, which is the behaviour
I want.

The ":after" kwarg needed to be modified for the keybinding, because
otherwise it wouldn't load properly (in fact, my previous binding was
shadowed by Doom's own function, but I never noticed).