Move `after!` blocks to calling sites for org
Instead of having to wrap the whole file in one big `after!` block, I
can instead wrap the call to load `org.el`. Then I can write the config
inside as normal, without having to worry about delayed evaluation.

Also moved the org-books config out into a separate file, in part to
adhere to the principle above, and in part due to a separation of
concerns.