pandoc: Fix punctuation bug temporarily
As per https://github.com/jgm/pandoc/issues/7509#issuecomment-1153842182
(somewhat badly formatted due to backticks), the issue of CJK fonts
highjacking punctuation in xelatex that has been bothering me for a
couple of years can be mitigated.

I am applying the fix at the default file level, meaning that every file
I compile with `-dpdf` will have it. After a couple of tests, it seems
to be working great: PDFs get compiled nicely with no fighting between
CJK fonts and smart punctuation.

However, the smart punctuation functionality is disabled by default for
org-mode input and has to be enabled with `--from=org+smart` when
compiling from the command line. The upside of this is that I no longer
need to use backticks when writing quotes in org-mode for later PDF
generation.