org: Add workaround for org-id links
ID links in org are not working properly.

When the `org-id-track-globally` var is set to true, the links are
stored in a hash table that is written to disk. However, Doom's config
sets `org-id-locations-file` to someplace in ~/org/, which is
exceedingly weird and probably a bug. This fix resets the file to its
default value.