Automatically recognize metadata files in pandoc
When converting from Emacs with `-dpdf`, the function will now look for
metadata files in the same directory and append them to the command if
they are present. I can use metadata files to adjust or enhance my
default settings this way.

I tried to do this within the default file itself, but couldn't get it
to work, so I'm using a programmatic approach for now.