Add fns to check for duplicate refs in org files
Pandoc does not automatically check for duplicate footnotes, and will
happily generate incorrect output.

According to my preliminary tests, it will warn on duplicate section
references, but still produce output, which is not great (maybe this is
customizable?).

Pandoc-crossref appears to block pandoc from compiling when there are
duplicate table references.

These functions run error checks before calling pandoc, and if any
errors occur, cancel the call. They currently shell out to awk and
friends, but perhaps could be rewritten in Elisp, if execution speed is
not an issue.