Automate CSL file choice when compiling w/ pandoc
Unlike some other options, pandoc does not allow `--csl` to be passed to
a command twice. If a defaults file includes a csl option, then it is
used even if another csl file is passed later in the command. Ergo,
defaults files do not work well for this.

I solved this by moving CSL choice directly into Elisp. The benefit here
is that I can use a single variable for different output formats. The
drawback is that it's less portable, and cannot be used in the command
line; although I pretty much never call pandoc from the CLI anyway.

Now my function will recognize a CSL file (only a single one) present in
the directory and prioritize it for citation styling. In the absence of
such a file, a default is provided.

I also moved the default file from Dropbox to the pandoc folder, so it's
synced together with my config.