pandoc-clj: refactor stretch filter
I have to return a Para from Clojure no matter what, even though it gets
deleted by pandoc later in the processing: an image with the "r-stretch"
class wrapped in a Para will get unwrapped automatically by pandoc.
Attempting to pass it a raw Image in JSON results in an error.