emacs: add fn to create link to file path
Sometimes I find myself in need of typing out the path to a file, which
is tedious without autocompletion. Opening the file with autocomplete
and yanking its path is a possibility, but it's a workaround (and not
always useful). Using dired is likewise a bit of an overkill.

Instead, I decided to leverage the existing autocomplete functionality
in Emacs (currently vertico) to find a file and insert its name at point
directly.