pandoc.clj: add nohandout filter, refactor
The `walk` lib is unnecessary, because the Pandoc AST is not really a
tree, but more of a list with occasional nested elements. So far all my
filters are working exclusively with top-level items, so a simple `map`
is enough.

The `nohandout` code deletes consecutive elements from the AST, so it
cannot be implemented as a map.